requirements.txt
setuptools
flask
pyyaml
python-telegram-bot
slackweb
matterhook